lodash added,git ignore updated,app.post /delete updated

+++ app.js
@@ -1,6 +1,7 @@
 const express = require("express");
 const bodyParser = require("body-parser");
 const mongoose = require("mongoose");
+const _ = require("lodash")
 const app = express();
 
 
@@ -37,7 +38,6 @@ const listSchema={
 const List=mongoose.model("List",listSchema);
 
 
-
 app.get("/",(req,res)=>{
    Item.find({},function(err,foundItems){
     if(foundItems.length===0){
@@ -57,7 +57,7 @@ app.get("/",(req,res)=>{
 })
 
 app.get("/:customListName",function(req,res){
-   const customListName= req.params.customListName
+   const customListName= _.capitalize(req.params.customListName);
 
    List.findOne({name:customListName},function(err,foundList){
     if(!err){
@@ -85,24 +85,47 @@ app.get("/:customListName",function(req,res){
 
 app.post("/",(req,res)=>{
     const itemName=req.body.newItem;
+    const listName=req.body.list; 
 
     const item = new Item({
         name:itemName
     });
-    item.save();
-    res.redirect("/");
+
+    if(listName === "Today"){
+        item.save();
+        res.redirect("/");
+    }else{
+        List.findOne({name:listName},(err,foundList)=>{
+           foundList.items.push(item);
+           foundList.save();
+           res.redirect("/"+listName);
+        });
+    }
+
     
-})
+    
+});
 
 app.post("/delete",(req,res)=>{
     const checkboxId= req.body.checkbox;
+    const listName= (req.body.listName);
 
-    Item.findByIdAndRemove(checkboxId,(err)=>{
-        if(!err){
-            console.log("Successfully removed item with id:"+checkboxId);
-        }
-        res.redirect("/")
-    })
+    if(listName==="Today"){
+        Item.findByIdAndRemove(checkboxId,(err)=>{
+            if(!err){
+                console.log("Successfully removed item with id:"+checkboxId);
+            }
+            res.redirect("/")
+        })
+    }else{
+        List.findOneAndUpdate({name:listName},{$pull:{items:{_id:checkboxId}}},function(err,foundList){
+            if(!err){
+                res.redirect("/"+listName);
+            }
+        })
+    }
+   
+   
 })
